requirements.txt
aioprometheus==22.5.0
ansible-core==2.14
decorator==5.1.1